refactor: modularize firebase api
