+++ src/lib/firebase/ratingsApi.js
@@ -1,0 +1,31 @@
+import { collection, collectionGroup, getDocs, deleteDoc, doc } from 'firebase/firestore';
+import baseApi from './baseApi.js';
+import { db } from '../firebase.js';
+import { errorHandler } from '../errorHandler.js';
+
+export const getBookRatings = (bookId) => baseApi.getCollection(`books/${bookId}/ratings`);
+export const addBookRating = (bookId, data) => baseApi.addToCollection(`books/${bookId}/ratings`, data);
+
+export async function getAllRatings() {
+  try {
+    const snap = await getDocs(collectionGroup(db, 'ratings'));
+    return snap.docs.map(d => ({ id: d.id, bookId: d.ref.parent.parent.id, ...d.data() }));
+  } catch (error) {
+    throw errorHandler.handleFirebaseError(error, 'ratings:get-all');
+  }
+}
+
+export async function deleteRating(bookId, ratingId) {
+  try {
+    await deleteDoc(doc(db, `books/${bookId}/ratings/${ratingId}`));
+  } catch (error) {
+    throw errorHandler.handleFirebaseError(error, `ratings:delete:${bookId}:${ratingId}`);
+  }
+}
+
+export default {
+  getBookRatings,
+  addBookRating,
+  getAllRatings,
+  deleteRating
+};
