+++ src/lib/services/order/orderItems.js
@@ -1,5 +1,5 @@
 import { errorHandler } from '../../errorHandler.js';
-import firebaseApi from '../../firebaseApi.js';
+import firebaseApi from '../../firebase/baseApi.js';
 
 export async function updateOrderItem(itemId, updateData) {
   try {
