+++ src/lib/services/order/refundPayment.js
@@ -1,5 +1,5 @@
 import { errorHandler } from '../../errorHandler.js';
-import firebaseApi from '../../firebaseApi.js';
+import firebaseApi from '../../firebase/baseApi.js';
 
 export async function refundOrderPayment(orderId, refundData) {
   try {
